\name{nifti.disp.lib.version}
\alias{nifti.disp.lib.version}
\title{NIfTI library version.}
\description{
Return NIfTI library version and date as a character string.
}
\usage{
nifti.disp.lib.version()
}
\value{
 String containing library version and compilation date.
}
\references{
  \url{http://nifti.nimh.nih.gov}
  \url{http://niftilib.sourceforge.net}
}
\seealso{ 
  \code{\link{nifti.image.new}},
  \code{\link{nifti.image.write}},
  \code{\link{nifti.image.read}}
}
\keyword{manip}
\keyword{data}

