\name{nifti_image_new}
\alias{nifti_image_new}
\title{Create new NIFTI image}
\description{
Basic initialization of a nifti.image structure (to a 1x1x1 image)
}
\usage{
nifti_image_new() 
}
\arguments{
}
\details{
}
\value{

}
\references{
\url{http://nifti.nimh.nih.gov}
\url{http://niftilib.sourceforge.net}
}
\author{
Oliver Granert <o.granert@neurologie.uni-kiel.de>
}

\seealso{ 
\code{\link{nifti_image_write} \link{nifti_image_read}}
}

\examples{
nifti_img=nifti_image_new()
}
\keyword{nifti_imgage_new}
